+++ src/components/level/toolbars/EditToolbar/EditToolbar.tsx
@@ -21,6 +21,7 @@ import { Button, ToolbarSeparator } from "ui/button";
 import { svgs } from "ui/icon";
 import { FindPlayerButton } from "./FindPlayerButton";
 import { SelectionEditButton } from "./SelectionEditButton";
+import { StatsButton } from "./StatsButton";
 import { TestingButtons } from "./TestingButtons";
 import { UndoButton } from "./UndoButton";
 
@@ -63,10 +64,10 @@ const HK_PASTE: HotKeyShortcuts = [["V", HotKeyMask.CTRL], ["Paste"]];
 const HK_DEL: HotKeyShortcuts = [["Delete"], ["Clear"]];
 
 interface Props {
-  withUndo?: boolean;
+  isCompact?: boolean;
 }
 
-export const EditToolbar: FC<Props> = ({ withUndo = true }) => {
+export const EditToolbar: FC<Props> = ({ isCompact = false }) => {
   const undoQueue = useStore($currentLevelUndoQueue)!;
   const noSelection = !useStore($hasSelection);
   const clipboardSize = useStore($clipboardRegionSizeStr);
@@ -95,7 +96,7 @@ export const EditToolbar: FC<Props> = ({ withUndo = true }) => {
 
   return (
     <>
-      {withUndo && <UndoButton />}
+      {isCompact || <UndoButton />}
       <Button
         icon={<svgs.Redo />}
         disabled={!undoQueue.canRedo}
@@ -131,8 +132,12 @@ export const EditToolbar: FC<Props> = ({ withUndo = true }) => {
         onClick={handleDelete}
         title={`Delete selection (${displayHotKey(HK_DEL)})`}
       />
-      <ToolbarSeparator />
-      <SelectionEditButton />
+      {isCompact || (
+        <>
+          <ToolbarSeparator />
+          <SelectionEditButton />
+        </>
+      )}
       {/*<Button disabled>PNG</Button>*/}
       <ToolbarSeparator />
       {/*<Button disabled>Copy level to clipboard</Button>*/}
@@ -141,6 +146,7 @@ export const EditToolbar: FC<Props> = ({ withUndo = true }) => {
       <TestingButtons />
       <ToolbarSeparator />
       <FindPlayerButton />
+      <StatsButton />
     </>
   );
 };