+++ src/components/level/toolbars/EditToolbar/selection-editors/rnd.tsx
@@ -1,21 +1,19 @@
 import cn from "classnames";
+import { createEvent, createStore } from "effector";
 import { useStore } from "effector-react";
-import { FC, useCallback, useMemo, useState } from "react";
+import { FC, useCallback, useMemo } from "react";
 import * as RoMap from "@cubux/readonly-map";
 import { getDriver, getTilesForToolbar } from "drivers";
 import { $currentDriverName } from "models/levelsets";
+import { HotKeyMask } from "models/ui/hotkeys";
 import { Button, TextButton } from "ui/button";
 import { svgs } from "ui/icon";
-import { Field, Select, SelectOption } from "ui/input";
+import { Field, Range } from "ui/input";
 import { ColorType } from "ui/types";
 import { EMPTY_MAP } from "utils/data";
 import { SelectionEditor, SelectionEditorProps } from "./_types";
 import clC from "./common.module.scss";
 import cl from "./rnd.module.scss";
-
-const countOptions = Array.from({ length: 31 }).map<SelectOption<number>>(
-  (_, i) => ({ value: i, label: String(i) }),
-);
 
 const getRandomTile = (
   prob: ReadonlyMap<number, number>,
@@ -31,6 +29,18 @@ const getRandomTile = (
   throw new Error("unreachable");
 };
 
+const toggleTile = createEvent<number>();
+const toggleTileKeep = toggleTile.prepend<any>(() => -1);
+const setCount = createEvent<[tile: number, count: number]>();
+const $probabilities = createStore<ReadonlyMap<number, number>>(EMPTY_MAP)
+  .on(toggleTile, (map, tile) =>
+    map.has(tile) ? RoMap.remove(map, tile) : RoMap.set(map, tile, 1),
+  )
+  .on(setCount, (map, [tile, count]) => RoMap.set(map, tile, count));
+const $total = $probabilities.map((prob) =>
+  RoMap.reduce(prob, (n, c) => n + c, 0),
+);
+
 const RndEditor: FC<SelectionEditorProps> = ({
   region,
   onSubmit,
@@ -43,8 +53,8 @@ const RndEditor: FC<SelectionEditorProps> = ({
     [region, tempLevelFromRegion],
   );
 
-  const [prob, setProb] = useState<ReadonlyMap<number, number>>(EMPTY_MAP);
-  const total = useMemo(() => RoMap.reduce(prob, (n, c) => n + c, 0), [prob]);
+  const prob = useStore($probabilities);
+  const total = useStore($total);
 
   const tilesSorted = useMemo(
     () =>
@@ -53,18 +63,7 @@ const RndEditor: FC<SelectionEditorProps> = ({
           ([, { value, metaTile }]) =>
             !metaTile || metaTile.primaryValue === value,
         )
-        .map(
-          ([, t]) =>
-            [
-              t,
-              () =>
-                setProb((map) =>
-                  map.has(t.value)
-                    ? RoMap.remove(map, t.value)
-                    : RoMap.set(map, t.value, 1),
-                ),
-            ] as const,
-        ),
+        .map(([, t]) => [t, () => toggleTile(t.value)] as const),
     [tiles],
   );
 
@@ -93,13 +92,7 @@ const RndEditor: FC<SelectionEditorProps> = ({
         <Button
           uiColor={prob.has(-1) ? ColorType.WARN : ColorType.MUTE}
           asLink={!prob.has(-1)}
-          onClick={useCallback(
-            () =>
-              setProb((map) =>
-                map.has(-1) ? RoMap.remove(map, -1) : RoMap.set(map, -1, 1),
-              ),
-            [],
-          )}
+          onClick={toggleTileKeep}
         >
           <i>keep old</i>
         </Button>
@@ -126,14 +119,11 @@ const RndEditor: FC<SelectionEditorProps> = ({
                   ) : (
                     <i>keep</i>
                   )}
-                  <Select
-                    options={countOptions}
-                    value={countOptions.find((o) => o.value === count) ?? null}
-                    onChange={(o) => {
-                      if (o) {
-                        setProb((map) => RoMap.set(map, tile, o.value));
-                      }
-                    }}
+                  <Range
+                    min={1}
+                    max={30}
+                    value={count}
+                    onChange={(v) => setCount([tile, v])}
                     className={cl.input}
                   />
                   <div className={cl.help}>
@@ -165,4 +155,5 @@ export const rnd: SelectionEditor = {
   title: "Random",
   icon: <svgs.Random />,
   Component: RndEditor,
+  hotkeys: ["N", HotKeyMask.SHIFT],
 };