+++ src/utils/rect/index.ts
@@ -2,4 +2,5 @@ export * from "./clipRect";
 export * from "./fromDrag";
 export * from "./inBounds";
 export * from "./inRect";
+export * from "./lineInterpolation";
 export * from "./types";